tests

+++ backend/models.py
@@ -1,12 +1,12 @@
 from database import Base
-from sqlalchemy import Column, Integer, String, TIMESTAMP,text, DECIMAL
+from sqlalchemy import Column, Integer, String, TIMESTAMP,text, Float
 
 class Order(Base):
     __tablename__ = "orders"
 
     id = Column(Integer,primary_key=True,nullable=False)
     symbol = Column(String, nullable=False)
-    price = Column(DECIMAL, nullable=False)
+    price = Column(Float, nullable=False)
     quantity = Column(Integer, nullable=False)
     order_type = Column(String, nullable=False)
     side = Column(String, nullable=False)
