+++ backend/schema.py
@@ -5,7 +5,7 @@ from typing_extensions import Annotated
 
 class OrderBase(BaseModel):
     symbol: str
-    price: Annotated[Decimal, Field(gt=0)]
+    price: Annotated[float, Field(gt=0)]
     quantity: Annotated[int, Field(strict=True,gt=0)]
     order_type: Literal["market", "limit", "stop"] = "market"
     side: Literal["buy","sell"]
